Use timestamp from property store when get prop.
Previously the timestamp in property events might come from two
different sources. For property update events, the timestamp is
generated from VehicleHalServer. For getProperty request, the
timestamp is generated from EmulatedVehicleHal. This might cause
a getProperty request which returns an older value to have a newer
timestamp compared to the newer property update event. As a result,
the new property update event will be ignored by the client, which
should not be the case since the property is actually updated and
the client should get the update.

This causes flaky test cases if the client is try to subscribe, which
will cause a get property request for the initial value and the client
is changing the value, which will cause a new property update event at
the same time.

Test: atest android.hardware.automotive.vehicle@2.0-default-impl-unit-tests
Bug: 311219830
Change-Id: I78613b1de30624011bd90a5d0da244f0e3d67337
(cherry picked from commit c69b857b8dc0457e66dda502e0adf089a692d516)