Remove the used_slots assertion from VtsHalWeaverTargetTest
Since fixing the boot-time issues in OMAPI based Weaver HALs has proven
to be more difficult than expected, for now the
config_disableWeaverOnUnsecuredUsers option is being added as a
workaround.  It weakens the security benefit of Weaver, but it's better
than not being able to enable Weaver at all.  As a result of supporting
this option, the !used_slots.empty() assertion has to be removed.

Bug: 368151356
Test: mmm hardware/interfaces/weaver
Change-Id: I2b06e477ee276296e2f8216c2ee99f4c526005a5