audio: Deprecate StreamSwitcher
Use of StreamSwitcher causes threading issues with the FMQ due to
using it in a way which is not anticipated by the FMQ design.

Stream implementations that need to change its behavior depending
on connected devices must do it without changing the I/O thread.
For example implementations please refer to PrimaryStream and
RemoteSubmixStream.

The class code has been moved to 'deprecated' directory and
namespace, and will be removed in the future. Any classes which
currently depend on StreamSwitcher should be updated mechanically
to use the new name and the include file location. Such classes
need to be planned to be overhauled as soon as possible to remove
the dependency on StreamSwitcher.

Bug: 300130515
Test: m
Change-Id: I6b0d20274013826360ca5efa69a01df9457db9c6