Fix size of Rect fields in CROP spec
The CROP spec specified to encode the rect as 4x int64, however
Rect, gralloc4's libgralloctypes, and mapper5's IMapperMetadataTypes
all used int32 for the Rect fields when encoding the crop since
that's the actual size of the Rect's fields.

Since VtsHalGraphicsMapperV4_0TargetTest's CROP test made use of
libgralloctypes, fix the size in the spec to align with the
implementations.

Fixes: 359690632
Test: N/A
Flag: EXEMPT doc only change
Change-Id: If829c31188bea0f203d6ea6a96695414e9513256