List required transitive deps of com.android.hardware.audio explicitly
These are prebuilt etc files which are installed outside this apex in
/vendor/etc/... Transitive deps that are installed outside the apex are
currently supported when building vendor.img with make, but the current
plan is that this behavior will not be ported when building vendor.img
with soong. To ensure that these deps are installed when building
vendor.img with soong, list these `required` deps explicitly on the top-level apex.

Test: verified that the list of installed files of vendor remain the
same
Bug: 374371755

Change-Id: I0961ab92d9db8e69ecb54b90ffd5b2511843c695