Expose more from hwtrust for DICE chain validation
There are a lot of upcoming requirements (see below) that place
restrictions on the CSR and the DICE chain. This is a first step to
make some of the infrastructure necessary to utilize
the already-existing functionality inside hwtrust.

Bug: 372843164
Bug: 376297620

A lot of parsing happens in `remote_prov_utils`. It would be nice to
have all parsing and validation occur in hwtrust.

Test: atest VtsHalRemotelyProvisionedComponentTargetTest
      atest libkeymint_remote_prov_support_test
Change-Id: I52383c0c62a6bb73554fd0771d2e7a8b366246c0