Fix formatting, use consistent comment styles, and document more fields.
Changes:
1) The aidl_format pre-upload hook complains about some line wrapping.

2) ASN.1 comments begin with "--". This only makes a difference in this case
if a vendor copy-pastes the schema in these comments somewhere that
expects valid a ASN.1 schema. It's also nice to be consistent.

As a drive-by improvement, I also added comments for all the fields in
KeyDescription. The statements about how attestationSecurityLevel and
keymintSecurityLevel must match are tested in VTS (and CTS).

Bug: 381383980
Test: n/a, comment update
Change-Id: I6a331fd3ed18e4a54ff8d34e96ae7dc62e3e3a0c