Use consistent, correct terminology for vendor API levels in VTS tests.
And make the RKP VTS tests refer to the SDK API levels consistently
(though it's nice to see that in a few places those tests refer to
the correct vendor API levels with the new YYYYMM format). I left
the YYYYMM references as-is since we'll do a proper refactor later
across all VTS tests, but used the __ANDROID_API_<foo>__ constants
everywhere else to be consistent with all the other VTS tests.

From what I can tell, there is no official notion of "VSR API level",
only "vendor API level". VSR refers to the Vendor Software Requirements,
so it makes sense to refer to "VSR-15" (for example) since that refers
to a specific version of the requirements. It therefore doesn't really
make sense to refer to "VSR-15 devices" (as opposed to "devices with vendor
API level 35" (in the same way that we don't refer to "CTS-15 devices", but
rather to "devices with SDK API level 35").

Bug: 384021501
Test: atest VtsAidlKeyMintTargetTest
Change-Id: I3914c26f81b3bdc886d06cc17f238b5de597ccfc