Convert SDK API level to vendor API level for Android 15+ in VTS gates.
This is a fairly targeted fix for b/383452093 since the new vendor API
level format (YYYYMM) was introduced in Android 14-QPR3 and therefore
only affects comparisons between the vendor API level and the SDK API
level starting with Android 15 (V).

A proper solution should not rely on the caller remembering to convert
the SDK API level, but this will do for now and we can improve the
gating properly in subsequent commits.

Bug: 383452093
Bug: 384021501
Test: atest VtsAidlKeyMintTargetTest
Change-Id: Ibb6ec2a294e96094ab5af8a674ddb77c1ad14993