Clarify requirement for ECDH functionality
The KeyMint HAL requires support for elliptic curves being used for ECDH
key agreement, but it's not mentioned in the headline documentation of
algorithms.

However, it's clear from the rest of the interface specification (e.g.
the KeyPurpose.aidl file), the CDD, and from the VTS/CTS tests that ECDH
support is required, so this is just a documentation clarification, not
a spec change.

Test: none, comment change
Change-Id: Ia1a11fefd6259b0d2ce1f1f15cb6d4b8a6b75089