bump.py: Bump Cuttlefish's target level and API level
This adds a step to bump.py that bumps Cuttlefish's SHIPPING_API_LEVEL
to the next SDK API level and bumps its target FCM level(both in trunk* and
next).
This should happen at the same time that we make the platform ready for
the bump so it's always tested.

This generates the same changes as https://r.android.com/3456364

Test: python bump.py 202504 202604 B C 16 17 36 37
Test: m
Bug: 389665041

Change-Id: I87719e220a8c713875937742ee6dad50450d5ada