Update annotation gen script.
Refactor annotation gen script to use formatter instead of header
and footer so that we can add AllowedAccessMode in addition to
DefaultAccessMode. This will be used in VTS test.

This refactor also includes a minor change to sort the annotations
so that the order is stable.

Flag: EXEMPT internal refactor
Test: atest VehiclePropertyAnnotationCppTest
Bug: 393209222

Change-Id: I21359a648dc6507c10f9fc9b8579f9639fd8d259